+++ src/PipServices.Commons.Netstandard16/Refer/Referencer.cs
@@ -12,11 +12,12 @@ namespace PipServices.Commons.Refer
         /// </summary>
         /// <param name="references">references to be assigned</param>
         /// <param name="components">a list of components to assign references</param>
-        public static void SetReferences(IReferences references, IEnumerable components)
+        public static void SetReferences(IReferences references, IEnumerable components = null)
         {
+            components = components ?? references.GetAll();
             foreach (var component in components)
             {
-                var referenceale = components as IReferenceable;
+                var referenceale = component as IReferenceable;
                 if (referenceale != null)
                     referenceale.SetReferences(references);
             }
@@ -35,5 +36,21 @@ namespace PipServices.Commons.Refer
                     unreferenceale.UnsetReferences();
             }
         }
+
+        /// <summary>
+        /// Clears references for components that implement IUnreferenceable interface 
+        /// </summary>
+        /// <param name="references">a list of components to clear references</param>
+        public static void UnsetReferences(IReferences references)
+        {
+            var components = references.GetAll();
+            foreach (var component in components)
+            {
+                var unreferenceale = component as IUnreferenceable;
+                if (unreferenceale != null)
+                    unreferenceale.UnsetReferences();
+            }
+        }
+
     }
 }